user_id - Update and Delete post with user_id

+++ app/routers/post.py
@@ -48,6 +48,11 @@ def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depe
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} does not exist")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action.")
+
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -64,6 +69,10 @@ def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} does not exist")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action.")
 
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
